# openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GemWorld B2B Gemstone Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory/{id}/boost:
    post:
      operationId: boostListing
      tags: [boosts]
      summary: Apply a boost to a gemstone listing
      description: Deducts one boost from the seller's available boosts and features the listing for 7 days
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BoostListingRequest"
      responses:
        "200":
          description: Listing boosted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoostResult"
        "400":
          description: No boosts available or invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Seller does not own this listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Listing or seller not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List all gemstones
      description: Returns all gemstones listed on GemWorld
      responses:
        "200":
          description: List of gemstones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Gemstone"
    post:
      operationId: addGemstone
      tags: [inventory]
      summary: Add a gemstone listing
      description: Seller adds a new gemstone to inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGemstoneRequest"
      responses:
        "201":
          description: Gemstone created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gemstone"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Seller not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Currency:
      type: string
      enum: [USD, INR, AED]
    Gemstone:
      type: object
      properties:
        id:
          type: string
        seller_id:
          type: string
        stone_type:
          type: string
        carat:
          type: number
          minimum: 0
        origin:
          type: string
        treatment:
          type: string
        price:
          type: number
          minimum: 0
        currency:
          $ref: "#/components/schemas/Currency"
        base_price_usd:
          type: number
          minimum: 0
        certificate_number:
          type: string
        image_url:
          type: string
        color:
          type: string
          nullable: true
          description: GIA colour grade (D–M). Required for diamonds.
        clarity:
          type: string
          nullable: true
          description: GIA clarity grade (IF, VVS1, VVS2, VS1, VS2, SI1, SI2, I1, I2, I3). Required for diamonds.
        rap_price_per_carat:
          type: number
          nullable: true
          description: "Rapaport Benchmark Price (USD/carat). Diamonds only."
        total_rap_value:
          type: number
          nullable: true
          description: "Total Rapaport value = rap_price_per_carat × carat. Diamonds only."
        is_featured:
          type: boolean
        boost_expiry_date:
          type: string
          format: date-time
          nullable: true
        estimated_price_min:
          type: number
          minimum: 0
          description: "Estimated Trade Range — floor (USD). Diamonds: 60% of Rap. Others: rule-based."
        estimated_price_max:
          type: number
          minimum: 0
          description: "Estimated Trade Range — ceiling (USD). Diamonds: 90% of Rap. Others: rule-based."
        pricing_confidence:
          type: string
          enum: [low, medium, high]
          description: Confidence level for the estimated price range
        pricing_disclaimer:
          type: string
          description: Disclaimer for the estimated price
        created_at:
          type: string
          format: date-time
      required:
        - id
        - seller_id
        - stone_type
        - carat
        - origin
        - treatment
        - price
        - currency
        - base_price_usd
        - certificate_number
        - image_url
        - is_featured
        - estimated_price_min
        - estimated_price_max
        - pricing_confidence
        - pricing_disclaimer
        - created_at
    CreateGemstoneRequest:
      type: object
      properties:
        seller_id:
          type: string
        stone_type:
          type: string
          minLength: 1
        carat:
          type: number
          minimum: 0
        origin:
          type: string
          minLength: 1
        treatment:
          type: string
          minLength: 1
        color:
          type: string
          description: GIA colour grade (D–M). Required when stone_type is diamond.
        clarity:
          type: string
          description: GIA clarity grade. Required when stone_type is diamond.
        price:
          type: number
          minimum: 0
        currency:
          $ref: "#/components/schemas/Currency"
        certificate_number:
          type: string
          minLength: 1
        image_url:
          type: string
          minLength: 1
      required:
        - seller_id
        - stone_type
        - carat
        - origin
        - treatment
        - price
        - currency
        - certificate_number
        - image_url
...
    BoostListingRequest:
      type: object
      properties:
        seller_id:
          type: string
      required:
        - seller_id
...
    BoostResult:
      type: object
      properties:
        gemstone_id:
          type: string
        is_featured:
          type: boolean
        boost_expiry_date:
          type: string
          format: date-time
        boosts_remaining:
          type: integer
        boost_cost_usd:
          type: number
        expires_in_days:
          type: integer
      required:
        - gemstone_id
        - is_featured
        - boost_expiry_date
        - boosts_remaining
        - boost_cost_usd
        - expires_in_days
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error